Added default to SENT column in Questions Table
changed the date_sent and received to text fields to store datetimes

Added jinja2 Email Template

moved create_session and initialised it. Also made 2 functions one to send emails to an individual address and the other to send them to a list.

Made a lot of changes to Mementos, but it currently works. Need to test time loops though